Add files via upload
This project helps a admin or a organization to maintain data of its employee. It has features like Add, update, remove or maintain a  data record set  of the employees that is a optimal solution for storing data and accessing it efficiently. 
languages used :    
1) Java : Java is a programming language that is widely used for developing a variety of applications, including web applications and etc. It is a class-based, object-oriented programming language.

2)Java Swing : It is a Java-based library used for developing graphical user interfaces (GUIs) for desktop applications
                      
3) JFrame :  which is used to create graphical user interfaces for desktop applications.

4)SQL(Structured Query Language): It  is a standardized programming language used to manage relational databases and perform various operations on the data in them.
Primary uses of SQL :
a)Data Retrieval
b)Data Manipulation
c)Data Definition 
and etc.